Factor out 2D overlay stuff into its own component.
